Create a general kuka control block. Torque control works on both simulation and hardware. Currently, the implemented controller is inverse dynamics controller + PD position controller. Next, we will target more general controllers. A new lcm message channel for controller reference is defined as lcmt_robot_controller_reference.
