+++ drake/examples/kuka_iiwa_arm/kuka_iiwa_signal_scope.py
@@ -3,11 +3,13 @@ import colorsys
 
 
 
-addPlot(timeWindow=15, yLimits=[-3.14, 3.14])
-addSignals('IIWA_STATUS', msg.utime, msg.joint_position_measured, range(7))
-addSignals('IIWA_STATUS', msg.utime, msg.joint_position_commanded, range(7))
+# addPlot(timeWindow=15, yLimits=[-3.14, 3.14])
+# addSignals('IIWA_STATUS', msg.utime, msg.joint_position_measured, range(7))
+# addSignals('IIWA_STATUS', msg.utime, msg.joint_position_commanded, range(7))
 
 
 addPlot(timeWindow=15, yLimits=[-3.14, 3.14])
 addSignals('IIWA_COMMAND', msg.utime, msg.joint_position, range(7))
 
+addPlot(timeWindow=15, yLimits=[-3.14, 3.14])
+addSignals('IIWA_COMMAND', msg.utime, msg.joint_torque, range(7))
