--- drake/examples/kuka_iiwa_arm/kuka_inverse_dynamics_runner.cc
@@ -1,171 +1,0 @@
-/// @file
-///
-/// kuka_inverse_dynamics_runner is designed to compute the torque command based on 
-/// desired joint position, velocity, and acceleration, and measured joint position and velocity.
-/// The final torque command is composed of inverse dynamics torque command and joint position PD
-/// controller command.
-/// Messages are sent via LCM channels.
-/// (TODO: Generalize this ID controller to more general types of feedback controllers)
-
-#include <lcm/lcm-cpp.hpp>
-
-#include "drake/common/drake_assert.h"
-#include "drake/common/drake_path.h"
-#include "drake/examples/kuka_iiwa_arm/iiwa_common.h"
-#include "drake/multibody/rigid_body_tree.h"
-
-#include "drake/lcmt_iiwa_inverse_dynamics_input.hpp"
-#include "drake/lcmt_iiwa_status.hpp"
-#include "drake/lcmt_iiwa_command.hpp"
-
-using Eigen::MatrixXd;
-using Eigen::VectorXd;
-using Eigen::VectorXi;
-using drake::Vector1d;
-using Eigen::Vector2d;
-using Eigen::Vector3d;
-
-namespace drake {
-namespace examples {
-namespace kuka_iiwa_arm {
-namespace {
-
-const char* const kLcmStatusChannel = "IIWA_STATUS";
-const char* const kLcmIDInputChannel = "IIWA_INVERSEDYNAMICS_INPUT";
-const char* const kLcmCommandChannel = "IIWA_COMMAND";
-const int kNumJoints = 7;
-
-class InverseDynamicsRunner {
- public:
-  /// tree is aliased
-  explicit InverseDynamicsRunner(const RigidBodyTree<double>& tree)
-      : tree_(tree), plan_number_(0), IDcontroller_trigger(false) {
-    VerifyIiwaTree(tree);
-    lcm_.subscribe(kLcmStatusChannel,
-                    &InverseDynamicsRunner::HandleStatus, this);
-    lcm_.subscribe(kLcmIDInputChannel,
-                    &InverseDynamicsRunner::HandleInverseDynamicsInput, this);
-  }
-
-  void Run() {
-    int64_t cur_time_us = -1;
-
-    // Initialize the timestamp to an invalid number so we can detect the first message.
-    iiwa_status_.utime = cur_time_us;
-    iiwa_inverse_dynamics_input_.utime = cur_time_us;
-
-    lcmt_iiwa_command iiwa_command;
-    iiwa_command.num_joints = kNumJoints;
-    iiwa_command.joint_position.resize(kNumJoints, 0.);
-    iiwa_command.num_torques = kNumJoints;
-    iiwa_command.joint_torque.resize(kNumJoints, 0.);
-
-    Eigen::VectorXd joint_position_desired(kNumJoints); 
-    Eigen::VectorXd joint_velocity_desired(kNumJoints); 
-    Eigen::VectorXd joint_accel_desired(kNumJoints); 
-
-    while (true) {
-      // Call lcm handle until at least one message is processed
-      while (0 == lcm_.handleTimeout(10)) { }
-      DRAKE_ASSERT(iiwa_status_.utime != -1);
-      cur_time_us = iiwa_status_.utime;
-
-      if (IDcontroller_trigger) {  
-        const int kNumDof = 7;
-        iiwa_command.utime = iiwa_status_.utime;
-
-        // Inverse Dynamics
-        // Set desired joint position, velocity and acceleration
-        for (int joint = 0; joint < kNumDof; joint++){
-          joint_position_desired(joint) = iiwa_inverse_dynamics_input_.joint_position_desired[joint];
-          joint_velocity_desired(joint) = iiwa_inverse_dynamics_input_.joint_velocity_desired[joint];
-          joint_accel_desired(joint) = iiwa_inverse_dynamics_input_.joint_accel_desired[joint];
-        }
-
-        double *qptr = &iiwa_status_.joint_position_measured[0];
-        Eigen::Map<Eigen::VectorXd> q(qptr, kNumDof);
-        double *qdptr = &iiwa_status_.joint_velocity_estimated[0];
-        Eigen::Map<Eigen::VectorXd> qd(qdptr, kNumDof);
-
-        // Computing inverse dynamics torque command
-        KinematicsCache<double> cache = tree_.doKinematics(q, qd);
-        const RigidBodyTree<double>::BodyToWrenchMap no_external_wrenches;
-
-        // note that, the gravity term in the inverse dynamics is set to zero.
-        Eigen::VectorXd torque_command = tree_.inverseDynamics(cache, no_external_wrenches, joint_accel_desired, false);
-        
-        Eigen::VectorXd z = Eigen::VectorXd::Zero(kNumDof); 
-        KinematicsCache<double> cache0 = tree_.doKinematics(q, qd);
-        Eigen::VectorXd gravity_torque = tree_.inverseDynamics(cache0, no_external_wrenches, z, false);
-        torque_command -= gravity_torque;
-
-        // PD position control
-        Eigen::VectorXd position_ctrl_torque_command(kNumDof);
-        Eigen::VectorXd Kp_pos_ctrl(kNumDof); // 7 joints
-        // Kp_pos_ctrl << 225, 289, 144, 49, 324, 49, 49;
-        Kp_pos_ctrl << 100, 100, 100, 100, 100, 50, 50;
-        Eigen::VectorXd Kd_pos_ctrl(kNumDof); // 7 joints
-        // Kd_pos_ctrl << 30, 34, 24, 14, 36, 14, 14;
-        Kd_pos_ctrl << 19, 19, 19, 19, 19, 14, 14;
-        // (TODOs) Add integral control (anti-windup)
-        for (int joint = 0; joint < kNumJoints; joint++) {
-          position_ctrl_torque_command(joint) = Kp_pos_ctrl(joint)*(joint_position_desired(joint) - iiwa_status_.joint_position_measured[joint])
-                                              + Kd_pos_ctrl(joint)*(joint_velocity_desired(joint) - iiwa_status_.joint_velocity_estimated[joint]);
-        }
-        //Combination of ID torque control and IK position control
-        torque_command += position_ctrl_torque_command;
-
-        // -------->(For Safety) Set up iiwa position command<----------
-        for (int joint = 0; joint < kNumJoints; joint++) {
-          iiwa_command.joint_position[joint] = joint_position_desired(joint);
-        }
-
-        // -------->Set up iiwa torque command<-------------
-        for (int joint = 0; joint < kNumJoints; joint++) {
-          iiwa_command.joint_torque[joint] = torque_command(joint);
-          iiwa_command.joint_torque[joint] = std::max(-150.0, std::min(150.0, iiwa_command.joint_torque[joint]));
-        }
-
-        lcm_.publish(kLcmCommandChannel, &iiwa_command);
-      }
-    }
-  }
-
- private:
-  void HandleStatus(const lcm::ReceiveBuffer* rbuf, const std::string& chan,
-                    const lcmt_iiwa_status* status) {
-    iiwa_status_ = *status;
-  }
-
-  void HandleInverseDynamicsInput(const lcm::ReceiveBuffer* rbuf, const std::string& chan,
-                    const lcmt_iiwa_inverse_dynamics_input* input) {
-    iiwa_inverse_dynamics_input_ = *input;
-    IDcontroller_trigger = true;
-  }
-
-  lcm::LCM lcm_;
-  const RigidBodyTree<double>& tree_;
-  int plan_number_{};
-  bool IDcontroller_trigger;// inverse dynamics runner wait for the first message from plan runner 
-  lcmt_iiwa_status iiwa_status_;
-  lcmt_iiwa_inverse_dynamics_input iiwa_inverse_dynamics_input_;
-};
-
-int do_main(int argc, const char* argv[]) {
-  RigidBodyTree<double> tree(
-      drake::GetDrakePath() + "/examples/kuka_iiwa_arm/urdf/iiwa14.urdf",
-      drake::multibody::joints::kFixed);
-
-  InverseDynamicsRunner runner(tree);
-  runner.Run();
-  return 0;
-}
-
-}  // namespace
-}  // namespace kuka_iiwa_arm
-}  // namespace examples
-}  // namespace drake
-
-int main(int argc, const char* argv[]) {
-  return drake::examples::kuka_iiwa_arm::do_main(argc, argv);
-}
