--- drake/lcmtypes/drake/lcmt_iiwa_inverse_dynamics_input.hpp
@@ -1,201 +1,0 @@
-/** THIS IS AN AUTOMATICALLY GENERATED FILE.  DO NOT MODIFY
- * BY HAND!!
- *
- * Generated by lcm-gen
- **/
-
-#include <lcm/lcm_coretypes.h>
-
-#ifndef __drake_lcmt_iiwa_inverse_dynamics_input_hpp__
-#define __drake_lcmt_iiwa_inverse_dynamics_input_hpp__
-
-#include <vector>
-
-namespace drake
-{
-
-/// Define a single set of joint states for the arm.
-class lcmt_iiwa_inverse_dynamics_input
-{
-    public:
-        int64_t    utime;
-
-        int32_t    num_joints;
-
-        std::vector< double > joint_position_desired;
-
-        std::vector< double > joint_velocity_desired;
-
-        std::vector< double > joint_accel_desired;
-
-    public:
-        /**
-         * Encode a message into binary form.
-         *
-         * @param buf The output buffer.
-         * @param offset Encoding starts at thie byte offset into @p buf.
-         * @param maxlen Maximum number of bytes to write.  This should generally be
-         *  equal to getEncodedSize().
-         * @return The number of bytes encoded, or <0 on error.
-         */
-        inline int encode(void *buf, int offset, int maxlen) const;
-
-        /**
-         * Check how many bytes are required to encode this message.
-         */
-        inline int getEncodedSize() const;
-
-        /**
-         * Decode a message from binary form into this instance.
-         *
-         * @param buf The buffer containing the encoded message.
-         * @param offset The byte offset into @p buf where the encoded message starts.
-         * @param maxlen The maximum number of bytes to reqad while decoding.
-         * @return The number of bytes decoded, or <0 if an error occured.
-         */
-        inline int decode(const void *buf, int offset, int maxlen);
-
-        /**
-         * Retrieve the 64-bit fingerprint identifying the structure of the message.
-         * Note that the fingerprint is the same for all instances of the same
-         * message type, and is a fingerprint on the message type definition, not on
-         * the message contents.
-         */
-        inline static int64_t getHash();
-
-        /**
-         * Returns "lcmt_iiwa_inverse_dynamics_input"
-         */
-        inline static const char* getTypeName();
-
-        // LCM support functions. Users should not call these
-        inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
-        inline int _getEncodedSizeNoHash() const;
-        inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
-};
-
-int lcmt_iiwa_inverse_dynamics_input::encode(void *buf, int offset, int maxlen) const
-{
-    int pos = 0, tlen;
-    int64_t hash = (int64_t)getHash();
-
-    tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
-    if(tlen < 0) return tlen; else pos += tlen;
-
-    tlen = this->_encodeNoHash(buf, offset + pos, maxlen - pos);
-    if (tlen < 0) return tlen; else pos += tlen;
-
-    return pos;
-}
-
-int lcmt_iiwa_inverse_dynamics_input::decode(const void *buf, int offset, int maxlen)
-{
-    int pos = 0, thislen;
-
-    int64_t msg_hash;
-    thislen = __int64_t_decode_array(buf, offset + pos, maxlen - pos, &msg_hash, 1);
-    if (thislen < 0) return thislen; else pos += thislen;
-    if (msg_hash != getHash()) return -1;
-
-    thislen = this->_decodeNoHash(buf, offset + pos, maxlen - pos);
-    if (thislen < 0) return thislen; else pos += thislen;
-
-    return pos;
-}
-
-int lcmt_iiwa_inverse_dynamics_input::getEncodedSize() const
-{
-    return 8 + _getEncodedSizeNoHash();
-}
-
-int64_t lcmt_iiwa_inverse_dynamics_input::getHash()
-{
-    static int64_t hash = _computeHash(NULL);
-    return hash;
-}
-
-const char* lcmt_iiwa_inverse_dynamics_input::getTypeName()
-{
-    return "lcmt_iiwa_inverse_dynamics_input";
-}
-
-int lcmt_iiwa_inverse_dynamics_input::_encodeNoHash(void *buf, int offset, int maxlen) const
-{
-    int pos = 0, tlen;
-
-    tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &this->utime, 1);
-    if(tlen < 0) return tlen; else pos += tlen;
-
-    tlen = __int32_t_encode_array(buf, offset + pos, maxlen - pos, &this->num_joints, 1);
-    if(tlen < 0) return tlen; else pos += tlen;
-
-    if(this->num_joints > 0) {
-        tlen = __double_encode_array(buf, offset + pos, maxlen - pos, &this->joint_position_desired[0], this->num_joints);
-        if(tlen < 0) return tlen; else pos += tlen;
-    }
-
-    if(this->num_joints > 0) {
-        tlen = __double_encode_array(buf, offset + pos, maxlen - pos, &this->joint_velocity_desired[0], this->num_joints);
-        if(tlen < 0) return tlen; else pos += tlen;
-    }
-
-    if(this->num_joints > 0) {
-        tlen = __double_encode_array(buf, offset + pos, maxlen - pos, &this->joint_accel_desired[0], this->num_joints);
-        if(tlen < 0) return tlen; else pos += tlen;
-    }
-
-    return pos;
-}
-
-int lcmt_iiwa_inverse_dynamics_input::_decodeNoHash(const void *buf, int offset, int maxlen)
-{
-    int pos = 0, tlen;
-
-    tlen = __int64_t_decode_array(buf, offset + pos, maxlen - pos, &this->utime, 1);
-    if(tlen < 0) return tlen; else pos += tlen;
-
-    tlen = __int32_t_decode_array(buf, offset + pos, maxlen - pos, &this->num_joints, 1);
-    if(tlen < 0) return tlen; else pos += tlen;
-
-    if(this->num_joints) {
-        this->joint_position_desired.resize(this->num_joints);
-        tlen = __double_decode_array(buf, offset + pos, maxlen - pos, &this->joint_position_desired[0], this->num_joints);
-        if(tlen < 0) return tlen; else pos += tlen;
-    }
-
-    if(this->num_joints) {
-        this->joint_velocity_desired.resize(this->num_joints);
-        tlen = __double_decode_array(buf, offset + pos, maxlen - pos, &this->joint_velocity_desired[0], this->num_joints);
-        if(tlen < 0) return tlen; else pos += tlen;
-    }
-
-    if(this->num_joints) {
-        this->joint_accel_desired.resize(this->num_joints);
-        tlen = __double_decode_array(buf, offset + pos, maxlen - pos, &this->joint_accel_desired[0], this->num_joints);
-        if(tlen < 0) return tlen; else pos += tlen;
-    }
-
-    return pos;
-}
-
-int lcmt_iiwa_inverse_dynamics_input::_getEncodedSizeNoHash() const
-{
-    int enc_size = 0;
-    enc_size += __int64_t_encoded_array_size(NULL, 1);
-    enc_size += __int32_t_encoded_array_size(NULL, 1);
-    enc_size += __double_encoded_array_size(NULL, this->num_joints);
-    enc_size += __double_encoded_array_size(NULL, this->num_joints);
-    enc_size += __double_encoded_array_size(NULL, this->num_joints);
-    return enc_size;
-}
-
-uint64_t lcmt_iiwa_inverse_dynamics_input::_computeHash(const __lcm_hash_ptr *)
-{
-    uint64_t hash = 0xc51a3fc3f2fbb508LL;
-    return (hash<<1) + ((hash>>63)&1);
-}
-
-}
-
-#endif
