--- drake/lcmtypes/lcmt_iiwa_inverse_dynamics_input.lcm
@@ -1,19 +1,0 @@
-package drake;
-
-// Define a single set of joint states for the arm.
-struct lcmt_iiwa_inverse_dynamics_input {
-  // The timestamp in microseconds.
-  int64_t utime;
-
-  // The number of joints. 
-  int32_t num_joints;
-
-  // The reference joint positions. 
-  double joint_position_desired[num_joints];
-
-  // The reference joint positions. 
-  double joint_velocity_desired[num_joints];
-
-  // The reference joint positions. 
-  double joint_accel_desired[num_joints];
-}
